+++ changelog.md
@@ -5,6 +5,11 @@
 ### Added
 
 -   Add `Highlight` component
+-   Add `SuggestInput` component
+
+### Removed
+
+-   Remove `Autocomplete` component
 
 ## [8.0.0-alpha.11] - 2019-06-07